Reinforcement Learning

Fill me in

Grid World

VLQuantitativeFinancePackage.MyPeriodicRectangularGridWorldModel — Type

mutable struct MyPeriodicRectangularGridWorldModel <: AbstractWorldModel

The MyPeriodicRectangularGridWorldModel mutable struct represents a rectangular grid world model.

Required fields

number_of_rows::Int: The number of rows in the grid

number_of_cols::Int: The number of columns in the grid

coordinates::Dict{Int,Tuple{Int,Int}}: A dictionary that holds the coordinates of each cell in the grid where the key is the cell index and the value is a tuple of the row and column index

states::Dict{Tuple{Int,Int},Int}: A dictionary that holds the states of each cell in the grid where the key is the cell coordinates and the value is the state index

moves::Dict{Int,Tuple{Int,Int}}: A dictionary that holds the moves that can be made from each cell in the grid where the key is the move index and the value is a tuple of the row and column changes associated with the move

rewards::Dict{Int,Float64}: A dictionary that holds the rewards for each cell in the grid where the key is the cell index and the value is the reward

VLQuantitativeFinancePackage.build — Method

function build(type::MyPeriodicRectangularGridWorldModel, nrows::Int, ncols::Int, 
    rewards::Dict{Tuple{Int,Int}, Float64}; defaultreward::Float64 = -1.0) -> MyPeriodicRectangularGridWorldModel

The build method constructs an instance of the MyPeriodicRectangularGridWorldModel type using the data in the NamedTuple.

Arguments

type::MyPeriodicRectangularGridWorldModel: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

nrows::Int: The number of rows in the grid

ncols::Int: The number of columns in the grid

rewards::Dict{Tuple{Int,Int}, Float64}: A dictionary that maps the coordinates of the grid to the rewards at those coordinates

defaultreward::Float64: The default reward for the grid. This is set to -1.0 by default.

Return

This function returns a populated instance of the MyPeriodicRectangularGridWorldModel type.

Wolfram policies

Fill me in

VLQuantitativeFinancePackage.MyOneDimensionalElementarWolframRuleModel — Type

mutable struct MyOneDimensionalElementarWolframRuleModel <: AbstractPolicyModel

The MyOneDimensionalElementarWolframRuleModel mutable struct represents a one-dimensional elementary Wolfram rule model.

Required fields

index::Int: The index of the rule

radius::Int: The radius, i.e, the number of cells that influence the next state for this rule

rule::Dict{Int,Int}: A dictionary that holds the rule where the key is the binary representation of the neighborhood and the value is the next state

VLQuantitativeFinancePackage.build — Method

function build(modeltype::Type{MyOneDimensionalElementarWolframRuleModel}, data::NamedTuple) -> MyOneDimensionalElementarWolframRuleModel

This build method constructs an instance of the MyOneDimensionalElementarWolframRuleModel type using the data in a NamedTuple.

Arguments

modeltype::Type{MyOneDimensionalElementarWolframRuleModel}: The type of model to build, in this case, the MyOneDimensionalElementarWolframRuleModel type.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

index::Int64: The index of the Wolfram rule

colors::Int64: The number of colors in the rule

radius::Int64: The radius, i.e., the number of cells to consider in the rule

Return

This function returns a populated instance of the MyOneDimensionalElementarWolframRuleModel type.