Data

We have included a market dataset that we use for examples and activities in the equity domain. This dataset holds the daily open, high, low, close, and volume data for a selection of stocks between 2014 and 2024.

VLQuantitativeFinancePackage.MyTrainingMarketDataSet — Function

MyTrainingMarketDataSet() -> Dict{String, DataFrame}

Load the components of the SP500 Daily open, high, low, close (OHLC) dataset as a dictionary of DataFrames. This data was provided by Polygon.io and covers the period from January 3, 2014, to December 31, 2024.

VLQuantitativeFinancePackage.MyTestingMarketDataSet — Function

MyTestingMarketDataSet() -> Dict{String, DataFrame}

Load the components of the SP500 Daily open, high, low, close (OHLC) dataset as a dictionary of DataFrames. This data was provided by Polygon.io and covers the period from January 3, 2025, to the current date (it is updated periodically).