Equity Securities

This should update now.

VLQuantitativeFinancePackage.EulerMaruyamaMethod — Type

struct EulerMaruyamaMethod <: AbstractStochasticSolverModel

Immutable type that represents the Euler-Maruyama method for solving stochastic differential equations. This type is passed to various functions in their method argument to indicate that the Euler-Maruyama method should be used in calculations.

VLQuantitativeFinancePackage.MyOrnsteinUhlenbeckModel — Type

mutable struct MyOrnsteinUhlenbeckModel <: AbstractAssetModel

A mutable struct that represents a Ornstein-Uhlenbeck process. An instance of MyOrnsteinUhlenbeckModel is configured and constructed using a corresponding build method.

Fields

μ::Function: The price drift function (long term price level)

σ::Function: The price volatility function

θ::Function: The mean reversion function. This function determines how quickly the price reverts to the long-term mean.

VLQuantitativeFinancePackage.MySisoLegSHippoModel — Type

mutable struct MySisoLegSHippoModel <: AbstractAssetModel

A mutable struct that represents a single-input, single-output (SISO) linear time-invariant (LTI) system that used the Leg-S parameterization. An instance of MySisoLegSHippoModel is configuired and constructed using a corresponding build method.

Fields

Â::Array{Float64,2}: Discretized state matrix of the system Â ∈ ℝ^(n x n) where n is the number of hidden states

B̂::Array{Float64,1}: Discretized input matrix of the system B̂ ∈ ℝ^n x 1

Ĉ::Array{Float64,1}: Discretized output matrix of the system Ĉ ∈ ℝ^1 x n

D̂::Array{Float64,1}: Discretized feedforward matrix of the system D̂ ∈ ℝ^1 x 1

n::Int: Number of hidden states in the system

Xₒ::Array{Float64,1}: Initial conditions of the system

VLQuantitativeFinancePackage.MyHestonModel — Type

mutable struct MyHestonModel <: AbstractAssetModel

A mutable struct that represents the Heston model. An instance of MyHestonModel is configured and constructed using a corresponding build method.

Fields

μ::Function: Drift function

κ::Function: Mean reversion function

θ::Function: Long-run volatility function

ξ::Function: Volatility of volatility function

Σ::Array{Float64,2}: Covariance matrix between the asset price and the volatility process

VLQuantitativeFinancePackage.solve — Function

solve(model::AbstractAssetModel, tspan::NamedTuple, initialconditions::AbstractArray; 
    method::AbstractStochasticSolverModel = EulerMaruyamaMethod(), N::Int64 = 100)::Tuple

The solve function is used to solve a stochastic asset model. The function takes a model, time span, initial conditions, method, and number of paths as input and returns a tuple of time and state arrays. The method argument is used to specify the method used to solve the model. The N argument is used to specify the number of paths to simulate. The function uses the Euler-Maruyama method as the default method to solve the model.

Arguments

model::AbstractAssetModel: The asset model to solve.

tspan::NamedTuple: The time span to solve the model over. The NamedTuple should have the fields start, stop, and step.

initialconditions::AbstractArray: The initial conditions for the model.

method::AbstractStochasticSolverModel = EulerMaruyamaMethod(): The method used to solve the model. Default is Euler-Maruyama.

N::Int64 = 100: The number of paths to simulate. Default is 100 paths.

Return

Tuple: A tuple of time and state arrays (more than one with multidimensional models) where the rows are the time steps and the columns are the paths.