Fixed Income Securities

Fixed income securities are financial instruments that pay a fixed amount of interest over a specified period of time. The most common fixed income securities are bonds, which are issued by governments, municipalities, and corporations. The fixed income market is one of the largest financial markets in the world, and it plays a critical role in the global economy.

Discounting models

In the VLQuantitativeFinancePackage we allow computing the present value of a cash flow stream using a discounting model. The present value of a cash flow stream can be computed using a ContinuousDiscountingModel or a DiscreteDiscountingModel.

VLQuantitativeFinancePackage.DiscreteCompoundingModel — Type

struct DiscreteCompoundingModel <: AbstractCompoundingModel

Immutable type that represents discrete compounding. This type has no fields and is passed as an argument to various functions to indicate that discrete compounding should be used in calculations.

VLQuantitativeFinancePackage.ContinuousCompoundingModel — Type

struct ContinuousCompoundingModel <: AbstractCompoundingModel

Immutable type that represents continuous compounding. This type has no fields and is passed as an argument to various functions to indicate that continuous compounding should be used in calculations.

Pricing models

VLQuantitativeFinancePackage.price — Function

price(model::MyUSTreasuryCouponSecurityModel, compounding::T) -> MyUSTreasuryCouponSecurityModel where T <: AbstractCompoundingModel

The price(...) function computes the price of a MyUSTreasuryCouponSecurityModel instance using a discrete or continuous compounding model.

Arguments

model::MyUSTreasuryCouponSecurityModel: an instance of the MyUSTreasuryCouponSecurityModel type.

compounding::T: an instance of a type that is a subtype of AbstractCompoundingModel, i.e., a discrete or continuous compounding model.

Returns

MyUSTreasuryCouponSecurityModel: an updated instance of the MyUSTreasuryCouponSecurityModel type with the price computed using the compounding model.

price(model::MyUSTreasuryZeroCouponBondModel, compounding::T) -> MyUSTreasuryZeroCouponBondModel where T <: AbstractCompoundingModel

The price(...) function computes the price of a MyUSTreasuryZeroCouponBondModel instance using a discrete or continuous compounding model.

Arguments

model::MyUSTreasuryZeroCouponBondModel: an instance of the MyUSTreasuryZeroCouponBondModel type.

compounding::T: an instance of a type that is a subtype of AbstractCompoundingModel, i.e., a discrete or continuous compounding model.

Returns

MyUSTreasuryZeroCouponBondModel: an updated instance of the MyUSTreasuryZeroCouponBondModel type with the price computed using the compounding model.

VLQuantitativeFinancePackage.strip — Function

strip(model::MyUSTreasuryCouponSecurityModel) -> Dict{Int, MyUSTreasuryZeroCouponBondModel}

Strips the coupon and par value payments from a parent coupon security.

The strip(...) function takes a model::MyUSTreasuryCouponSecurityModel of the security we wish to strip and returns a Dictionary holding MyUSTreasuryZeroCouponBondModel instances created from the parent security. The keys of the dictionary correspond to the temporal index of the created security.