Portfolio management

Fill me in

VLQuantitativeFinancePackage.MyMarkowitzRiskyAssetOnlyPortfiolioChoiceProblem — Type

mutable struct MyMarkowitzRiskyAssetOnlyPortfiolioChoiceProblem <: AbstractStochasticChoiceProblem

The MyMarkowitzRiskyAssetOnlyPortfiolioChoiceProblem mutable struct represents a Minimum Variance portfolio problem with risky assets only.

Required fields

Σ::Array{Float64,2}: The covariance matrix of the risky asset Returns

μ::Array{Float64,1}: The expected returns of the risky assets

bounds::Array{Float64,2}: The bounds on the risky asset weights

R::Float64: The desired return of the portfolio

initial::Array{Float64,1}: The initial portfolio weights

VLQuantitativeFinancePackage.MyMarkowitzRiskyRiskFreePortfiolioChoiceProblem — Type

mutable struct MyMarkowitzRiskyRiskFreePortfiolioChoiceProblem <: AbstractStochasticChoiceProblem

The MyMarkowitzRiskyRiskFreePortfiolioChoiceProblem mutable struct represents a Minimum Variance portfolio problem with a combination of risky and risk-free assets.

Required fields

Σ::Array{Float64,2}: The covariance matrix of the risky asset returns

μ::Array{Float64,1}: The expected returns of the risky assets

bounds::Array{Float64,2}: The bounds on the risky asset weights

R::Float64: The desired return of the portfolio

initial::Array{Float64,1}: The initial portfolio weights

risk_free_rate::Float64: The risk-free rate of return

VLQuantitativeFinancePackage.MySingleIndexModel — Type

mutable struct MySingleIndexModel <: AbstractReturnModel

The MySingleIndexModel mutable struct represents a single index model of risky asset returns.

Required fields

α::Float64: The firm specific unexplained return

β::Float64: The relationship between the firm and the market

r::Float64: The risk-free rate of return

ϵ::Distribution: The random shocks to the model (unexplained return)

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyMarkowitzRiskyAssetOnlyPortfiolioChoiceProblem}, data::NamedTuple)  -> MyMarkowitzRiskyAssetOnlyPortfiolioChoiceProblem

This build method constructs an instance of the MyMarkowitzRiskyAssetOnlyPortfiolioChoiceProblem type using the data in a NamedTuple.

Arguments

model::Type{MyMarkowitzRiskyAssetOnlyPortfiolioChoiceProblem}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

μ::Array{Float64,1}: The drift rates of the assets.

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyMarkowitzRiskyRiskFreePortfiolioChoiceProblem}, data::NamedTuple) -> MyMarkowitzRiskyRiskFreePortfiolioChoiceProblem

This build method constructs an instance of the MyMarkowitzRiskyRiskFreePortfiolioChoiceProblem type using the data in a NamedTuple.

Arguments

model::Type{MyMarkowitzRiskyRiskFreePortfiolioChoiceProblem}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

Σ::Array{Float64,2}: The covariance matrix of the risky asset returns

μ::Array{Float64,1}: The expected returns of the risky assets

bounds::Array{Float64,2}: The bounds on the risky asset weights

R::Float64: The desired return of the portfolio

initial::Array{Float64,1}: The initial portfolio weights

risk_free_rate::Float64: The risk-free rate of return

VLQuantitativeFinancePackage.build — Method

build(model::Type{MySingleIndexModel}, data::NamedTuple) -> MySingleIndexModel

This build method constructs an instance of the MySingleIndexModel type using the data in a NamedTuple.

Arguments

model::Type{MySingleIndexModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

α::Float64: The firm specific unexplained return

β::Float64: The relationship between the firm and the market

r::Float64: The risk-free rate of return

ϵ::Distribution: The random shocks to the model (unexplained return)

Return

This function returns an instance of the MySingleIndexModel type.

See also:

The unexplained return is modeled as a random variable, and the user can specify the distribution of this variable using the ϵ::Distribution argument.

See the Distributions.jl package package for more information on the available distributions.