Derivative Securities

Fill me in

European and American Options Contracts

VLQuantitativeFinancePackage.MyEuropeanCallContractModel — Type

mutable struct MyEuropeanCallContractModel <: AbstractContractModel

The MyEuropeanCallContractModel mutable struct represents a European call option contract. A call option is a financial contract that gives the holder the right, but not the obligation, to buy an asset at a specified price (the strike price) within a specified time period. For European options, the contract can only be exercised by the buyer at the expiration date.

Required fields

K::Float64: The strike price of the option.

Optional fields

sense::Union{Nothing, Int64}: The sense of the option. A value of 1 indicates a long position, and a value of -1 indicates a short position.

DTE::Union{Nothing,Float64}: The days to expiration of the option (measured in units of years).

IV::Union{Nothing, Float64}: The implied volatility of the option contract.

premium::Union{Nothing, Float64}: The premium of the option contract. This is the price paid to the seller by the buyer for the option contract.

ticker::Union{Nothing,String}: The ticker symbol of the underlying asset.

copy::Union{Nothing, Int64}: Number of contracts purchased or sold.

VLQuantitativeFinancePackage.MyEuropeanPutContractModel — Type

mutable struct MyEuropeanPutContractModel <: AbstractContractModel

The MyEuropeanPutContractModel mutable struct represents a European put option contract. A put option is a financial contract that gives the holder the right, but not the obligation, to sell an asset at a specified price (the strike price) within a specified time period. For European options, the contract can only be exercised by the buyer at the expiration date.

Required fields

K::Float64: The strike price of the option.

Optional fields

sense::Union{Nothing, Int64}: The sense of the option. A value of 1 indicates a long position, and a value of -1 indicates a short position.

DTE::Union{Nothing,Float64}: The days to expiration of the option (measured in units of years).

IV::Union{Nothing, Float64}: The implied volatility of the option contract.

premium::Union{Nothing, Float64}: The premium of the option contract. This is the price paid to the seller by the buyer for the option contract.

ticker::Union{Nothing,String}: The ticker symbol of the underlying asset.

copy::Union{Nothing, Int64}: Number of contracts purchased or sold.

VLQuantitativeFinancePackage.MyAmericanCallContractModel — Type

mutable struct MyAmericanCallContractModel <: AbstractContractModel

The MyAmericanCallContractModel mutable struct represents an American call option contract. An American call option is a financial contract that gives the holder the right, but not the obligation, to buy an asset at a specified price (the strike price). American option contracts can be exercised at any time on or before the expiration date.

Required fields

K::Float64: The strike price of the option.

Optional fields

sense::Union{Nothing, Int64}: The sense of the option. A value of 1 indicates a long position, and a value of -1 indicates a short position.

DTE::Union{Nothing,Float64}: The days to expiration of the option (measured in units of years).

IV::Union{Nothing, Float64}: The implied volatility of the option contract.

premium::Union{Nothing, Float64}: The premium of the option contract. This is the price paid to the seller by the buyer for the option contract.

ticker::Union{Nothing,String}: The ticker symbol of the underlying asset.

copy::Union{Nothing, Int64}: Number of contracts purchased or sold.

VLQuantitativeFinancePackage.MyAmericanPutContractModel — Type

mutable struct MyAmericanPutContractModel <: AbstractContractModel

The MyAmericanPutContractModel mutable struct represents an American put option contract. A put option is a financial contract that gives the holder the right, but not the obligation, to sell an asset at a specified price (the strike price). For American options, the contract can be exercised at any time on or before the expiration date.

Required fields

K::Float64: The strike price of the option.

Optional fields

sense::Union{Nothing, Int64}: The sense of the option. A value of 1 indicates a long position, and a value of -1 indicates a short position.

DTE::Union{Nothing,Float64}: The days to expiration of the option (measured in units of years).

IV::Union{Nothing, Float64}: The implied volatility of the option contract.

premium::Union{Nothing, Float64}: The premium of the option contract. This is the price paid to the seller by the buyer for the option contract.

ticker::Union{Nothing,String}: The ticker symbol of the underlying asset.

copy::Union{Nothing, Int64}: Number of contracts purchased or sold.

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyEuropeanCallContractModel}, data::NamedTuple) -> MyEuropeanCallContractModel

This build method constructs an instance of the MyEuropeanCallContractModel type using the data in a NamedTuple.

Arguments

model::Type{MyEuropeanCallContractModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

K::Float64: The strike price of the contract.

The other fields of the MyEuropeanCallContractModel model are set to nothing by default. These optional fields can be updated by the user after the model is built, or by passing the values in the data::NamedTuple argument.

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyEuropeanPutContractModel}, data::NamedTuple) -> MyEuropeanPutContractModel

This build method constructs an instance of the MyEuropeanPutContractModel type using the data in a NamedTuple.

Arguments

model::Type{MyEuropeanPutContractModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

K::Float64: The strike price of the contract.

The other fields of the MyEuropeanPutContractModel model are set to nothing by default. These optional fields can be updated by the user after the model is built, or by passing the values in the data::NamedTuple argument.

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyAmericanCallContractModel}, data::NamedTuple) -> MyAmericanCallContractModel

This build method constructs an instance of the MyAmericanCallContractModel type using the data in a NamedTuple.

Arguments

model::Type{MyAmericanCallContractModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

K::Float64: The strike price of the contract.

The other fields of the MyAmericanCallContractModel model are set to nothing by default. These optional fields can be updated by the user after the model is built, or by passing the values in the data::NamedTuple argument.

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyAmericanPutContractModel}, data::NamedTuple) -> MyAmericanPutContractModel

This build method constructs an instance of the MyAmericanPutContractModel type using the data in a NamedTuple.

Arguments

model::Type{MyAmericanPutContractModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

K::Float64: The strike price of the contract.

The other fields of the MyAmericanPutContractModel model are set to nothing by default. These optional fields can be updated by the user after the model is built, or by passing the values in the data::NamedTuple argument.

Payoff and profit of contracts at expiration

VLQuantitativeFinancePackage.payoff — Function

payoff(contracts::Array{T,1}, S::Array{Float64,1}) -> Array{Float64,2} where T <: AbstractContractModel

The payoff function computes the payoff for a set of option contracts at expiration given the underlying prices contained in the S::Array{Float64,1} array.

Arguments

contracts::Array{T,1}: An array of option contracts where T is a subtype of the AbstractContractModel type.

S::Array{Float64,1}: An array of underlying prices.

Returns

Array{Float64,2}: A matrix of size (number_of_underlying_prices, number_of_contracts + 2) where each row represents an underlying price and each column containts the payoff for each contract.

The first column contains the underlying price, the second column contains the payoff for the first contract, the third column contains the payoff for the second contract, and so on. The last column contains the sum of the payoffs for all contracts.

VLQuantitativeFinancePackage.profit — Function

profit(contracts::Array{T,1}, S::Array{Float64,1}) -> Array{Float64,2} where T <: AbstractContractModel

The profit function computes the profit for a set of option contracts at expiration given the underlying prices contained in the S::Array{Float64,1} array. This function requires the contracts to have the premium field set on each contract model.

Arguments

contracts::Array{T,1}: An array of option contracts where T is a subtype of the AbstractContractModel type. Each contract must have the premium field set.

S::Array{Float64,1}: An array of underlying prices at expiration.

Returns

Array{Float64,2}: A matrix of size (number_of_underlying_prices, number_of_contracts + 2) where each row represents an underlying price and each column containts the profit for a contract.

The first column contains the underlying price, the second column contains the profit for the first contract, the third column contains the profit for the second contract, and so on. The last column contains the sum of the profits for all contracts.

Computing European contract premiums

VLQuantitativeFinancePackage.MyBlackScholesContractPricingModel — Type

mutable struct MyBlackScholesContractPricingModel <: AbstractAssetModel

The MyBlackScholesContractPricingModel mutable struct represents a Black-Scholes model for pricing European option contracts.

Required fields

r::Float64: The annual risk-free discount rate

Sₒ::Float64: The current price of the underlying asset

VLQuantitativeFinancePackage.premium — Method

premium(contract::MyEuropeanCallContractModel, 
    model::MyBlackScholesContractPricingModel; sigdigits::Int64 = 4)

The premium function computes the premium for a European call option contract using the Black-Scholes-Merton model. This function requires the contract to have the K, DTE, and IV fields set on the contract model, and the Sₒ and r fields set on the pricing model::MyBlackScholesContractPricingModel instance.

Arguments

contract::MyEuropeanCallContractModel: An instance of the MyEuropeanCallContractModel type which models the European call option contract.

model::MyBlackScholesContractPricingModel: An instance of the MyBlackScholesContractPricingModel type which models the Black-Scholes-Merton model calculation.

Returns

Float64: The premium for the European call option contract.

VLQuantitativeFinancePackage.premium — Method

premium(contract::MyEuropeanPutContractModel, 
    model::MyBlackScholesContractPricingModel; sigdigits::Int64 = 4) -> Float64

The premium function computes the premium for a European put option contract using the Black-Scholes-Merton model. This function requires the contract to have the K, DTE, and IV fields set on the contract model, and the Sₒ and r fields set on the pricing model::MyBlackScholesContractPricingModel instance.

Arguments

contract::MyEuropeanPutContractModel: An instance of the MyEuropeanPutContractModel type which models the European put option contract.

model::MyBlackScholesContractPricingModel: An instance of the MyBlackScholesContractPricingModel type which models the Black-Scholes-Merton model calculation.

Returns

Float64: The premium for the European put option contract.

Computing American contract premiums

VLQuantitativeFinancePackage.MyAdjacencyBasedCRREquityPriceTree — Type

mutable struct MyAdjacencyBasedCRREquityPriceTree <: AbstractEquityPriceTreeModel

The MyAdjacencyBasedCRREquityPriceTree mutable struct represents a Cox-Ross-Rubinstein (CRR) model for pricing American style option contracts. The lattice is constructed using an adjacency list to represent the connectivity of the nodes in the lattice. The lattice is populated using the populate function, and the contract pricing is performed using the premium function.

Required fields

μ::Float64: The drift rate of the asset price

σ::Float64: The volatility of the asset price

T::Float64: The time to expiration of the option contract (measured in units of years)

Optional or computed fields

ΔT::Union{Nothing,Float64}: The time step size

u::Union{Nothing,Float64}: The up-factor for the lattice

d::Union{Nothing,Float64}: The down-factor for the lattice

p::Union{Nothing,Float64}: The probability of an up move in the lattice

data::Union{Nothing, Dict{Int, MyCRRLatticeNodeModel}}: A dictionary that holds the lattice data for each node where nodes are modeled as MyCRRLatticeNodeModel instances.

connectivity::Union{Nothing, Dict{Int64, Array{Int64,1}}}: A dictionary that holds the connectivity of the lattice where the key is the node index and the value is an array of the connected nodes.

levels::Union{Nothing, Dict{Int64,Array{Int64,1}}}: A dictionary that holds the nodes on each level of the lattice where the key is the level index and the value is an array of the nodes on that level.

The optional fields are computed when the lattice is passed to the populate function.

VLQuantitativeFinancePackage.MyCRRLatticeNodeModel — Type

mutable struct MyCRRLatticeNodeModel

The MyCRRLatticeNodeModel mutable struct represents a node in a Cox-Ross-Rubinstein (CRR) model for pricing American style option contracts.

Required fields

price::Float64: The price of the price at the node

probability::Float64: The probability of reaching the node

Optional or computed fields

intrinsic::Union{Nothing,Float64}: The intrinsic value of the option contract at the node

extrinsic::Union{Nothing,Float64}: The extrinsic value of the option contract at the node

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyAdjacencyBasedCRREquityPriceTree}, data::NamedTuple) -> MyAdjacencyBasedCRREquityPriceTree

This build method constructs an instance of the MyAdjacencyBasedCRREquityPriceTree type using the data in a NamedTuple.

Arguments

model::Type{MyAdjacencyBasedCRREquityPriceTree}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

μ::Float64: The drift rate of the asset price. For a risk-neutral measure, this is the risk-free rate.

σ::Float64: The volatility of the asset price. This is the implied volatility for a risk-neutral measure.

T::Float64: The time to expiration of the option contract (measured in units of years)

VLQuantitativeFinancePackage.populate — Method

populate(model::MyAdjacencyBasedCRREquityPriceTree; 
    Sₒ::Float64 = 100.0, h::Int64 = 1)

The populate function initializes the MyAdjacencyBasedCRREquityPriceTree model with share prices and probabilities for each node in the lattice. In addition, this methods sets the intrinsic and extrinsic values of each node to 0.0 and computes the connectivity and levels of the lattice.

Arguments

model::MyAdjacencyBasedCRREquityPriceTree: An instance of the MyAdjacencyBasedCRREquityPriceTree type.

Sₒ::Float64 = 100.0: The initial price of the equity.

h::Int64 = 1: The number of time steps in the lattice (the height of the binomial price tree).

Returns

MyAdjacencyBasedCRREquityPriceTree: An updated instance of the MyAdjacencyBasedCRREquityPriceTree type with the share prices and probabilities computed for each node in the lattice.

VLQuantitativeFinancePackage.premium — Function

premium(contract::T, model::MyAdjacencyBasedCRREquityPriceTree; 
    choice::Function=_rational, sigdigits::Int64 = 4) -> Float64 where {T<:AbstractContractModel}

Computes the premium for an American style option contract using the Cox-Ross-Rubinstein model.

Arguments

contract::T: An instance of the AbstractContractModel type which models the option contract.

model::MyAdjacencyBasedCRREquityPriceTree: An instance of the MyAdjacencyBasedCRREquityPriceTree type which models the Cox-Ross-Rubinstein model.

choice::Function=_rational: A function that determines the choice of the option contract. Default value is _rational.

sigdigits::Int64 = 4: The number of significant digits to round the premium to. Default value is 4.

Returns

Float64: The premium for the American style option contract.

premium(contract::MyEuropeanCallContractModel, 
    model::MyBlackScholesContractPricingModel; sigdigits::Int64 = 4)

The premium function computes the premium for a European call option contract using the Black-Scholes-Merton model. This function requires the contract to have the K, DTE, and IV fields set on the contract model, and the Sₒ and r fields set on the pricing model::MyBlackScholesContractPricingModel instance.

Arguments

contract::MyEuropeanCallContractModel: An instance of the MyEuropeanCallContractModel type which models the European call option contract.

model::MyBlackScholesContractPricingModel: An instance of the MyBlackScholesContractPricingModel type which models the Black-Scholes-Merton model calculation.

Returns

Float64: The premium for the European call option contract.

premium(contract::MyEuropeanPutContractModel, 
    model::MyBlackScholesContractPricingModel; sigdigits::Int64 = 4) -> Float64

The premium function computes the premium for a European put option contract using the Black-Scholes-Merton model. This function requires the contract to have the K, DTE, and IV fields set on the contract model, and the Sₒ and r fields set on the pricing model::MyBlackScholesContractPricingModel instance.

Arguments

contract::MyEuropeanPutContractModel: An instance of the MyEuropeanPutContractModel type which models the European put option contract.

model::MyBlackScholesContractPricingModel: An instance of the MyBlackScholesContractPricingModel type which models the Black-Scholes-Merton model calculation.

Returns

Float64: The premium for the European put option contract.