Factory

We provide a factory function build to construct most of the types used in this package.

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyUSTreasuryZeroCouponBondModel}, data::NamedTuple) -> MyUSTreasuryZeroCouponBondModel

This build method constructs an instance of the MyUSTreasuryZeroCouponBondModel type using the data in the NamedTuple argument.

Arguments

model::Type{MyUSTreasuryZeroCouponBondModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple argument must contain the following keys:

par::Float64: The face or par value of the bond

rate::Union{Nothing, Float64}: Effective annual interest rate (discount rate specified as a decimal)

T::Union{Nothing,Float64}: Duration in years measured as a 365 day or a 52 week year

n::Int: Number of compounding periods per year (typically 2)

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyUSTreasuryCouponSecurityModel}, data::NamedTuple) -> MyUSTreasuryCouponSecurityModel

This build method constructs an instance of the MyUSTreasuryCouponSecurityModel type using the data in the NamedTuple.

Arguments

model::Type{MyUSTreasuryCouponSecurityModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

par::Float64: Par value of the bond

rate::Union{Nothing, Float64}: Annualized effective discount rate

coupon::Union{Nothing, Float64}: Coupon interest rate

T::Union{Nothing,Float64}: Duration in years of the note or bond, measured as a 365 day or a 52 week year

λ::Int: Number of coupon payments per year (typically 2)

Return

This function returns an instance of the MyUSTreasuryCouponSecurityModel type. To populate the model, use the price function or a short-cut function involving a compounding model.

Example

Let's build a MyUSTreasuryCouponSecurityModel instance and populate the price, cashflow and discount datastructures for a T = 20-yr bond, with a coupon rate of c = 1.750%, a yield (discount rate) rate = 1.850%, two coupon payments per year, i.e., $\lambda = 2$ and a face (par) value of $V_{P}$ = 100 USD:

test_bond = build(MyUSTreasuryCouponSecurityModel, (
    T = 20.0, rate = 0.01850, coupon = 0.01750, λ = 2, par = 100.0
)) |> discount_model;

where the discount_model refers to either a DiscreteCompoundingModel or a ContinuousCompoundingModel instance.

VLQuantitativeFinancePackage.build — Method

build(modeltype::Type{MyOrnsteinUhlenbeckModel}, data::NamedTuple) -> MyOrnsteinUhlenbeckModel

This method builds an instance of the MyOrnsteinUhlenbeckModel type using the data in a NamedTuple.

Arguments

modeltype::Type{MyOrnsteinUhlenbeckModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple argument must contain the following keys:

μ::Float64: The long-term mean of the process.

σ::Float64: The volatility of the process.

θ::Float64: The mean reversion rate of the process.

VLQuantitativeFinancePackage.build — Method

build(modeltype::Type{MySisoLegSHippoModel}, data::NamedTuple) -> MySisoLegSHippoModel

This build method constructs an instance of the MySisoLegSHippoModel type using the data in a NamedTuple. This implementation uses the bilinear method to discretize the model, where the A and B matrices are computed using the Leg-S parameterization.

Arguments

modeltype::Type{MySisoLegSHippoModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

number_of_hidden_states::Int64: The number of hidden states in the model.

Δt::Float64: The time step size used to discretize the model (constant).

uₒ::Array{Float64,1}: The initial input to the model.

C::Array{Float64,1}: The output matrix of the model.

VLQuantitativeFinancePackage.build — Method

build(model::Type{MyHestonModel}, data::NamedTuple) -> MyHestonModel

This build method constructs an instance of the MyHestonModel type using the data in a NamedTuple.

Arguments

model::Type{MyHestonModel}: The type of model to build.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

μ::Function: The long-term mean of the process.

κ::Function: The mean reversion rate of the process.

θ::Function: The volatility of the process.

ξ::Function: The volatility of the volatility of the process.

Σ::Array{Float64,2}: The covariance matrix of the process.