Fixed Income Securities

Fixed income securities are financial instruments that pay a fixed amount of interest over a specified period of time. The most common fixed income securities are bonds, which are issued by governments, municipalities, and corporations. The fixed income market is one of the largest financial markets in the world, and it plays a critical role in the global economy.

Discounting models

In the VLQuantitativeFinancePackage we allow computing the present value of a cash flow stream using a discounting model. The present value of a cash flow stream can be computed using a ContinuousDiscountingModel or a DiscreteDiscountingModel.

VLQuantitativeFinancePackage.DiscreteCompoundingModel — Type

struct DiscreteCompoundingModel <: AbstractCompoundingModel

Immutable type that represents discrete compounding. This type has no fields and is passed as an argument to various functions to indicate that discrete compounding should be used in calculations.

VLQuantitativeFinancePackage.ContinuousCompoundingModel — Type

struct ContinuousCompoundingModel <: AbstractCompoundingModel

Immutable type that represents continuous compounding. This type has no fields and is passed as an argument to various functions to indicate that continuous compounding should be used in calculations.

Pricing models

We model United States Treasury securities, e.g., Treasury bills, Treasury notes, and Treasury bonds using the MyUSTreasuryZeroCouponBondModel and MyUSTreasuryCouponSecurityModel types, which are subtypes of the AbstractTreasuryDebtSecurity abstract type.

VLQuantitativeFinancePackage.MyUSTreasuryZeroCouponBondModel — Type

mutable struct MyUSTreasuryZeroCouponBondModel <: AbstractTreasuryDebtSecurity

A mutable struct that represents a U.S. Treasury zero coupon bond.

Fields

par::Float64: Par value of the bond

rate::Union{Nothing, Float64}: Annual interest rate

T::Union{Nothing,Float64}: Duration in years, measured as a 365 day or a 52 week year

price::Union{Nothing, Float64}: Price of the bond or note

n::Int: Number of compounding periods per year (typically 2)

cashflow::Union{Nothing, Dict{Int,Float64}}: Cashflow dictionary where the key is the period and the value is the discounted cashflow in a period

discount::Union{Nothing, Dict{Int,Float64}}: Discount factor dictionary where the key is the period and the value is the discount factor in that period

VLQuantitativeFinancePackage.MyUSTreasuryCouponSecurityModel — Type

mutable struct MyUSTreasuryCouponSecurityModel <: AbstractTreasuryDebtSecurity

A mutable struct that represents a U.S. Treasury coupon bond. This type of security (note or bond) pays the holder of the note (or bond) a fixed interest rate at regular intervals over the life of the instrument.

Fields

par::Float64: Par value of the bond

rate::Union{Nothing, Float64}: Annualized effective discount rate

coupon::Union{Nothing, Float64}: Coupon interest rate

T::Union{Nothing,Float64}: Duration in years of the note or bond, measured as a 365 day or a 52 week year

λ::Int: Number of coupon payments per year (typically 2)

price::Union{Nothing, Float64}: Price of the bond or note

cashflow::Union{Nothing, Dict{Int,Float64}}: Cashflow dictionary where the key is the period and the value is the discounted cashflow in a period

discount::Union{Nothing, Dict{Int,Float64}}: Discount factor dictionary where the key is the period and the value is the discount factor in that period

We construct MyUSTreasuryZeroCouponBondModel and MyUSTreasuryCouponSecurityModel objects by specifying the maturity date, the face value of the bond the effective annual interest rate and other data that are specific to the bill, note or bond in a build function where the data is specified in a NamedTuple format.

For example, let's compute the price and cashflow for a T = 20-yr bond, with a coupon rate of c = 1.750%, a yield (discount rate) rate = 1.850%, two coupon payments per year, i.e., $\lambda = 2$ and a face (par) value of $V_{P}$ = 100 USD

test_bond = build(MyUSTreasuryCouponSecurityModel, (
    T = 20.0, rate = 0.01850, coupon = 0.01750, λ = 2, par = 100.0
)) |> discount_model;

VLQuantitativeFinancePackage.price — Function

price(model::MyUSTreasuryCouponSecurityModel, compounding::T) -> MyUSTreasuryCouponSecurityModel where T <: AbstractCompoundingModel

The price(...) function computes the price of a MyUSTreasuryCouponSecurityModel instance using a discrete or continuous compounding model.

Arguments

model::MyUSTreasuryCouponSecurityModel: an instance of the MyUSTreasuryCouponSecurityModel type.

compounding::T: an instance of a type that is a subtype of AbstractCompoundingModel, i.e., a discrete or continuous compounding model.

Returns

MyUSTreasuryCouponSecurityModel: an updated instance of the MyUSTreasuryCouponSecurityModel type with the price computed using the compounding model.

price(model::MyUSTreasuryZeroCouponBondModel, compounding::T) -> MyUSTreasuryZeroCouponBondModel where T <: AbstractCompoundingModel

The price(...) function computes the price of a MyUSTreasuryZeroCouponBondModel instance using a discrete or continuous compounding model.

Arguments

model::MyUSTreasuryZeroCouponBondModel: an instance of the MyUSTreasuryZeroCouponBondModel type.

compounding::T: an instance of a type that is a subtype of AbstractCompoundingModel, i.e., a discrete or continuous compounding model.

Returns

MyUSTreasuryZeroCouponBondModel: an updated instance of the MyUSTreasuryZeroCouponBondModel type with the price computed using the compounding model.

VLQuantitativeFinancePackage.strip — Function

strip(model::MyUSTreasuryCouponSecurityModel) -> Dict{Int, MyUSTreasuryZeroCouponBondModel}

Strips the coupon and par value payments from a parent coupon security.

The strip(...) function takes a model::MyUSTreasuryCouponSecurityModel of the security we wish to strip and returns a Dictionary holding MyUSTreasuryZeroCouponBondModel instances created from the parent security. The keys of the dictionary correspond to the temporal index of the created security.