Fixed Income Treasury Securities

Fixed income securities are financial instruments that pay a fixed amount of interest over a specified period of time. The most common fixed income securities are bonds, which are issued by governments, municipalities, and corporations. The fixed income market is one of the largest financial markets in the world, and it plays a critical role in the global economy.

Discounting moddel

In the VLQuantitativeFinancePackage we allow computing the present value of a cash flow stream using a discounting model. The present value of a cash flow stream can be computed using a ContinuousCompoundingModel or a DiscreteCompoundingModel.

VLQuantitativeFinancePackage.DiscreteCompoundingModel — Type

struct DiscreteCompoundingModel <: AbstractCompoundingModel

Immutable type that represents discrete compounding. This type has no fields and is passed as an argument to various functions to indicate that discrete compounding should be used in calculations.

VLQuantitativeFinancePackage.ContinuousCompoundingModel — Type

struct ContinuousCompoundingModel <: AbstractCompoundingModel

Immutable type that represents continuous compounding. This type has no fields and is passed as an argument to various functions to indicate that continuous compounding should be used in calculations.

Treasury secruity model types

We model United States Treasury debt securities, e.g., Treasury bills, Treasury notes, and Treasury bonds using the MyUSTreasuryZeroCouponBondModel and MyUSTreasuryCouponSecurityModel types, which are subtypes of the AbstractTreasuryDebtSecurity abstract type.

VLQuantitativeFinancePackage.MyUSTreasuryZeroCouponBondModel — Type

mutable struct MyUSTreasuryZeroCouponBondModel <: AbstractTreasuryDebtSecurity

A mutable struct that represents a U.S. Treasury zero coupon security.

Fields

par::Float64: The face or par value of the bond

rate::Union{Nothing, Float64}: Effective annual interest rate (discount rate specified as a decimal)

T::Union{Nothing,Float64}: Duration in years of the instrument, measured as a 365 day or a 52 week year

n::Int: Number of compounding periods per year (typically 2)

Computed Fields

price::Union{Nothing, Float64}: Price of the bond or note (computed property)

cashflow::Union{Nothing, Dict{Int,Float64}}: Cashflow dictionary (computed property) where the key is the period and the value is the discounted cashflow in a period

discount::Union{Nothing, Dict{Int,Float64}}: Discount factor dictionary (computed property) where the key is the period and the value is the discount factor in that period

VLQuantitativeFinancePackage.MyUSTreasuryCouponSecurityModel — Type

mutable struct MyUSTreasuryCouponSecurityModel <: AbstractTreasuryDebtSecurity

A mutable struct that represents a U.S. Treasury coupon bond. This type of security (note or bond) pays the holder of the instrument a fixed interest rate at regular intervals over the life of the security.

Fields

par::Float64: Par value of the bond

rate::Union{Nothing, Float64}: Annualized effective discount rate

coupon::Union{Nothing, Float64}: Coupon interest rate

T::Union{Nothing,Float64}: Duration in years of the note or bond, measured as a 365 day or a 52 week year

λ::Int: Number of coupon payments per year (typically 2)

Computed Fields

price::Union{Nothing, Float64}: Price of the bond or note (computed property)

cashflow::Union{Nothing, Dict{Int,Float64}}: Cashflow dictionary (computed property) where the key is the period and the value is the discounted cashflow in a period

discount::Union{Nothing, Dict{Int,Float64}}: Discount factor dictionary(computed property) where the key is the period and the value is the discount factor in that period

Computing Treasury security prices

To compute the price of a Treasury security, we use the price function that takes a AbstractTreasuryDebtSecurity object and a discounting model as input arguments. We construct MyUSTreasuryZeroCouponBondModel and MyUSTreasuryCouponSecurityModel objects by specifying the maturity date, the face value of the bond the effective annual interest rate and other data that are specific to the bill, note or bond in a build function where the data is specified in a NamedTuple format.

VLQuantitativeFinancePackage.price — Function

price(model::MyUSTreasuryCouponSecurityModel, compounding::T) -> MyUSTreasuryCouponSecurityModel where T <: AbstractCompoundingModel

The price(...) function computes the price of a MyUSTreasuryCouponSecurityModel instance using a discrete or continuous compounding model.

Arguments

model::MyUSTreasuryCouponSecurityModel: an instance of the MyUSTreasuryCouponSecurityModel type.

compounding::T: an instance of a type that is a subtype of AbstractCompoundingModel, i.e., a discrete or continuous compounding model.

Returns

MyUSTreasuryCouponSecurityModel: an updated instance of the MyUSTreasuryCouponSecurityModel type with the price computed using the compounding model.

price(model::MyUSTreasuryZeroCouponBondModel, compounding::T) -> MyUSTreasuryZeroCouponBondModel where T <: AbstractCompoundingModel

The price(...) function computes the price of a MyUSTreasuryZeroCouponBondModel instance using a discrete or continuous compounding model.

Arguments

model::MyUSTreasuryZeroCouponBondModel: an instance of the MyUSTreasuryZeroCouponBondModel type.

compounding::T: an instance of a type that is a subtype of AbstractCompoundingModel, i.e., a discrete or continuous compounding model.

Returns

MyUSTreasuryZeroCouponBondModel: an updated instance of the MyUSTreasuryZeroCouponBondModel type with the price computed using the compounding model.

Short-cut syntax

We also provide a short-cut syntax to compute the price of a Treasury security which indirectly calls the price function. The short-cut syntax allows the user to specify the data required to construct a AbstractTreasuryDebtSecurity object, and then pass the data to the price function in a single call.

The short-cut syntax is as follows:

(compounding::DiscreteCompoundingModel)(model::MyUSTreasuryCouponSecurityModel) = _price_discrete_compounding(model::MyUSTreasuryCouponSecurityModel)
(compounding::ContinuousCompoundingModel)(model::MyUSTreasuryCouponSecurityModel) = _price_continuous_compounding(model::MyUSTreasuryCouponSecurityModel)
(compounding::DiscreteCompoundingModel)(model::MyUSTreasuryZeroCouponBondModel) = _price_discrete_compounding(model::MyUSTreasuryZeroCouponBondModel)
(compounding::ContinuousCompoundingModel)(model::MyUSTreasuryZeroCouponBondModel) = _price_continuous_compounding(model::MyUSTreasuryZeroCouponBondModel)

To use the short-cut syntax, the user must first construct a AbstractTreasuryDebtSecurity object using the build function and then pass the object to the short-cut syntax function. For example, to compute the price of a 20-year Treasury bond with a coupon rate of 1.750%, a yield (discount rate) of 1.850%, two coupon payments per year, and a face value of 100 USD, the user can use the following code:

test_bond = build(MyUSTreasuryCouponSecurityModel, (
    T = 20.0, rate = 0.01850, coupon = 0.01750, λ = 2, par = 100.0
)) |> discount_model;

where the discount_model refers to either a DiscreteCompoundingModel or a ContinuousCompoundingModel instance.

Separating the principal and interest payments

Registered Interest and Principal of Securities (STRIPS) bonds are a unique type of fixed-income investment instrument that provides investors with an alternative way to access the coupon payments of Treasury securities. STRIPS bonds are created by separating a Treasury securities coupon and principal components and trading them as individual zero-coupon securities. This process allows investors to purchase and trade the coupon or principal components separately, providing greater flexibility in managing their investment portfolios.

For example, a 5-year Treasury note with annual coupon payments of $C$ USD and a face (par) value of $V_{P}$ (USD) can be stripped into six separate zero-coupon securities, i.e., five zero-coupon bonds, each with face values of $C$ and maturity of $T$= 1,2,3,4 and 5 years, and a six security with face (par) value of $V_{P}$ USD with a duration of $T$ = 5 years. In the general case, a treasury note or bond with $N=\lambda{T}$ coupon payments, where $T$ denotes the maturity in years, and $\lambda$ represents the number of coupon payments per year, can be stripped into $N+1$ separate zero-coupon securities.

VLQuantitativeFinancePackage.strip — Function

strip(model::MyUSTreasuryCouponSecurityModel) -> Dict{Int, MyUSTreasuryZeroCouponBondModel}

Strips the coupon and par value payments from a parent coupon security.

The strip(...) function takes a model::MyUSTreasuryCouponSecurityModel of the security we wish to strip and returns a Dictionary holding MyUSTreasuryZeroCouponBondModel instances created from the parent security. The keys of the dictionary correspond to the temporal index of the created security.